+++ packages/components/checkbox/src/CheckboxGroupItem.tsx
@@ -29,32 +29,29 @@ const CheckboxGroupItem = forwardRef<HTMLInputElement, CheckboxGroupItemProps>(
     const { children, helpText, value, disabled: disabledProp } = props;
 
     const interactionProps = useInteractions({ disabled: disabledProp });
-    const store = useCheckboxGroupContext();
+    const { canSelectAll, hasError, orientation, state } =
+      useCheckboxGroupContext();
     const internalRef = useRef<HTMLInputElement>(null);
-    const mergedRef = useMergeRefs(internalRef, ref);
-    const { inputProps } = useCheckboxGroupItem(
-      props,
-      store,
-      mergedRef as unknown as React.RefObject<HTMLInputElement>
-    );
-    const disabled = store.isDisabled || props.disabled;
-    const isSelected = store.isSelected(value);
+    const mergeRefs = useMergeRefs(internalRef, ref);
+    const { inputProps } = useCheckboxGroupItem(props, state, internalRef);
+    const disabled = state.isDisabled || props.disabled;
+    const isSelected = state.isSelected(value);
 
     const dynamicStyles = useMemo((): FlexProps => {
-      const maybeCanSelectAll = store.canSelectAll ? canSelectAllStyles : {};
+      const maybeCanSelectAll = canSelectAll ? canSelectAllStyles : {};
 
-      if (store.orientation === 'vertical') {
+      if (orientation === 'vertical') {
         return {
           ...maybeCanSelectAll,
           _notLast: { marginBottom: 5 },
-          _notFirst: { marginLeft: store.canSelectAll ? 11 : 4 },
+          _notFirst: { marginLeft: canSelectAll ? 11 : 4 },
         };
       }
       return {
         ...maybeCanSelectAll,
         _notLast: { marginRight: 5 },
       };
-    }, [store.orientation, store.canSelectAll]);
+    }, [orientation, canSelectAll]);
 
     return (
       <CheckboxWrapper
@@ -64,13 +61,13 @@ const CheckboxGroupItem = forwardRef<HTMLInputElement, CheckboxGroupItemProps>(
         {...dynamicStyles}
         {...interactionProps}>
         <VisuallyHidden>
-          <input {...inputProps} ref={mergedRef} />
+          <input {...inputProps} ref={mergeRefs} />
         </VisuallyHidden>
         <CheckboxControl
           disabled={disabled}
           isSelected={isSelected}
           indeterminate={props.indeterminate}
-          isInvalid={store?.hasError}
+          isInvalid={hasError}
           {...interactionProps}
         />
         <CheckboxLabel disabled={disabled}>{children}</CheckboxLabel>