+++ packages/components/menu/src/MenuItem/MenuItem.tsx
@@ -1,6 +1,8 @@
 import React, { forwardRef } from 'react';
 
+import { MenuButtonPrimitive } from '@real-system/menu-primitive';
 import { MenuItemPrimitive } from '@real-system/menu-primitive';
+import { type ColorSchemes } from '@real-system/styled-library';
 import { makeTestId } from '@real-system/utils-library';
 
 import type { CommonMenuItemProps } from './MenuItem.model';
@@ -11,7 +13,10 @@ import { MenuItemIcon } from './MenuItemIcon';
 import { MenuItemLink } from './MenuItemLink';
 import { MenuItemRadio } from './MenuItemRadio';
 
-type MenuItemProps = CommonMenuItemProps;
+type MenuItemProps = CommonMenuItemProps & {
+  colorScheme?: ColorSchemes;
+  isSubmenu?: boolean;
+};
 
 export interface MenuItemComponent
   extends React.ForwardRefExoticComponent<MenuItemProps> {
@@ -24,12 +29,16 @@ export interface MenuItemComponent
 
 // @ts-ignore MenuItem properties are defined below
 const MenuItem: MenuItemComponent = forwardRef<HTMLDivElement, MenuItemProps>(
-  function MenuItem({ children, ...restProps }, ref) {
+  function MenuItem({ children, isSubmenu, ...restProps }, ref) {
     return (
       <MenuItemPrimitive
-        render={<MenuItemWrapper />}
+        render={
+          <MenuItemWrapper
+            as={isSubmenu && MenuButtonPrimitive}
+            {...restProps}
+          />
+        }
         data-testid={makeTestId('menu-item')}
-        {...restProps}
         ref={ref}>
         {children}
       </MenuItemPrimitive>