+++ packages/components/menu/src/MenuItem/MenuItemIcon.tsx
@@ -4,31 +4,21 @@ import type { IconProps } from '@real-system/icon';
 import { Icon } from '@real-system/icon';
 import { makeTestId } from '@real-system/utils-library';
 
-type MenuItemIconProps = (
-  | {
-      alignLeft?: boolean;
-      alignRight?: never;
-    }
-  | {
-      alignLeft?: never;
-      alignRight?: boolean;
-    }
-) &
-  Omit<IconProps, 'size'>;
+type MenuItemIconProps = {
+  alignRight?: boolean;
+} & Omit<IconProps, 'size'>;
 
 const MenuItemIcon = forwardRef<HTMLSpanElement, MenuItemIconProps>(
-  function MenuItemIcon({ alignRight, alignLeft = true, ...restProps }, ref) {
-    const spaceProps = alignLeft
-      ? { marginRight: 5 }
-      : alignRight
+  function MenuItemIcon({ alignRight, ...restProps }, ref) {
+    const spaceProps = alignRight
       ? { marginLeft: 'auto', marginRight: 0 }
-      : {};
+      : { marginRight: 5 };
 
     return (
       <Icon
         data-testid={makeTestId('menu-item-icon')}
-        {...spaceProps}
         {...restProps}
+        {...spaceProps}
         ref={ref}
       />
     );