+++ packages/components/menu/stories/Menu.stories.tsx
@@ -20,6 +20,7 @@ export const Default = () => (
           Edit
           <Menu.Item.Command>E</Menu.Item.Command>
         </Menu.Item>
+
         <Menu.Item>
           <Menu.Item.Icon icon="share" />
           Share
@@ -42,6 +43,63 @@ export const Default = () => (
           showExternal>
           Report issue
         </Menu.Item.Link>
+      </Menu.List>
+    </Menu>
+  </Box>
+);
+
+export const Submenu = () => (
+  <Box height="24rem">
+    <Menu>
+      <Menu.Button variant="fill">Actions</Menu.Button>
+      <Menu.List>
+        <Menu.Item>
+          <Menu.Item.Icon icon="pencil-alt" />
+          Edit
+          <Menu.Item.Command>E</Menu.Item.Command>
+        </Menu.Item>
+
+        <Menu.Item>
+          <Menu.Item.Icon icon="share" />
+          Share
+          <Menu.Item.Command>S</Menu.Item.Command>
+        </Menu.Item>
+        <Menu.Item>
+          <Menu.Item.Icon icon="archive" />
+          Archive
+          <Menu.Item.Command>A</Menu.Item.Command>
+        </Menu.Item>
+        <Menu.Item disabled>
+          <Menu.Item.Icon icon="trash" />
+          Delete
+          <Menu.Item.Command>D</Menu.Item.Command>
+        </Menu.Item>
+        <Menu.Separator />
+        <Menu placement="right-start">
+          <Menu.Button>Get Support</Menu.Button>
+          <Menu.List>
+            <Menu.Item.Link
+              href="https://system.themikewolf.com"
+              external
+              showExternal>
+              Report issue
+            </Menu.Item.Link>
+            <Menu.Item>
+              Chat Support
+              <Menu.Item.Icon icon="chat" alignRight />
+            </Menu.Item>
+            <Menu placement="right-start">
+              <Menu.Button>Other Options</Menu.Button>
+              <Menu.List>
+                <Menu.Item>
+                  <Menu.Item.Icon icon="trash" />
+                  Delete
+                  <Menu.Item.Command>E</Menu.Item.Command>
+                </Menu.Item>
+              </Menu.List>
+            </Menu>
+          </Menu.List>
+        </Menu>
       </Menu.List>
     </Menu>
   </Box>