+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @real-system/dialog
 
+## 0.0.58
+
+### Patch Changes
+
+- [#1128](https://github.com/bigwoof91/real-system/pull/1128) [`9992fc4`](https://github.com/bigwoof91/real-system/commit/9992fc4bf0d86a35ab4e11b82776a501952f6c86) Thanks [@bigwoof91](https://github.com/bigwoof91)! - upgrade ariakit, react-aria and fix breaking api's
+
+- Updated dependencies [[`9992fc4`](https://github.com/bigwoof91/real-system/commit/9992fc4bf0d86a35ab4e11b82776a501952f6c86)]:
+  - @real-system/elements-primitive@0.0.27
+  - @real-system/animation-library@0.0.25
+  - @real-system/dialog-primitive@0.0.36
+  - @real-system/styled-library@0.0.25
+  - @real-system/state-library@0.0.24
+  - @real-system/utils-library@0.0.37
+  - @real-system/typography@0.0.64
+  - @real-system/button@0.0.66
+  - @real-system/flex@0.0.56
+
+## 0.0.57
+
+### Patch Changes
+
+- Updated dependencies [[`5803f943`](https://github.com/bigwoof91/real-system/commit/5803f94371f08912a2bef2513d9783675630efcc), [`aa445777`](https://github.com/bigwoof91/real-system/commit/aa445777ff680c8b03006de72d217b2df8fdfbe7)]:
+  - @real-system/styled-library@0.0.24
+  - @real-system/utils-library@0.0.36
+  - @real-system/button@0.0.65
+  - @real-system/flex@0.0.55
+  - @real-system/typography@0.0.63
+  - @real-system/elements-primitive@0.0.26
+
 ## 0.0.56
 
 ### Patch Changes