+++ packages/components/radio-group/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @real-system/radio-group
 
+## 0.0.44
+
+### Patch Changes
+
+- [#1128](https://github.com/bigwoof91/real-system/pull/1128) [`9992fc4`](https://github.com/bigwoof91/real-system/commit/9992fc4bf0d86a35ab4e11b82776a501952f6c86) Thanks [@bigwoof91](https://github.com/bigwoof91)! - upgrade ariakit, react-aria and fix breaking api's
+
+- Updated dependencies [[`9992fc4`](https://github.com/bigwoof91/real-system/commit/9992fc4bf0d86a35ab4e11b82776a501952f6c86)]:
+  - @real-system/elements-primitive@0.0.27
+  - @real-system/visually-hidden@0.0.59
+  - @real-system/ariakit-library@0.0.28
+  - @real-system/styled-library@0.0.25
+  - @real-system/state-library@0.0.24
+  - @real-system/utils-library@0.0.37
+  - @real-system/a11y-library@0.0.20
+  - @real-system/typography@0.0.64
+  - @real-system/flex@0.0.56
+
+## 0.0.43
+
+### Patch Changes
+
+- [#1090](https://github.com/bigwoof91/real-system/pull/1090) [`aa445777`](https://github.com/bigwoof91/real-system/commit/aa445777ff680c8b03006de72d217b2df8fdfbe7) Thanks [@bigwoof91](https://github.com/bigwoof91)! - remove unused useMergeRefs and add truthy check to useReadyRef
+
+- Updated dependencies [[`5803f943`](https://github.com/bigwoof91/real-system/commit/5803f94371f08912a2bef2513d9783675630efcc), [`aa445777`](https://github.com/bigwoof91/real-system/commit/aa445777ff680c8b03006de72d217b2df8fdfbe7)]:
+  - @real-system/ariakit-library@0.0.27
+  - @real-system/styled-library@0.0.24
+  - @real-system/utils-library@0.0.36
+  - @real-system/flex@0.0.55
+  - @real-system/typography@0.0.63
+  - @real-system/visually-hidden@0.0.58
+  - @real-system/elements-primitive@0.0.26
+
 ## 0.0.42
 
 ### Patch Changes