+++ packages/libraries/descendants-library/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @real-system/descendants-library
 
+## 0.0.20
+
+### Patch Changes
+
+- [#1128](https://github.com/bigwoof91/real-system/pull/1128) [`9992fc4`](https://github.com/bigwoof91/real-system/commit/9992fc4bf0d86a35ab4e11b82776a501952f6c86) Thanks [@bigwoof91](https://github.com/bigwoof91)! - upgrade ariakit, react-aria and fix breaking api's
+
+- Updated dependencies [[`9992fc4`](https://github.com/bigwoof91/real-system/commit/9992fc4bf0d86a35ab4e11b82776a501952f6c86)]:
+  - @real-system/state-library@0.0.24
+  - @real-system/utils-library@0.0.37
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [[`aa445777`](https://github.com/bigwoof91/real-system/commit/aa445777ff680c8b03006de72d217b2df8fdfbe7)]:
+  - @real-system/utils-library@0.0.36
+
 ## 0.0.18
 
 ### Patch Changes